# More efficient parsing of documents