Before:
```
[info] Benchmark                      (size)   Mode  Cnt         Score        Error  Units
[info] GoogleMapsAPIReading.smithy4s     N/A  thrpt    5     12944.819 ±    187.492  ops/s
[info] NestedStructsReading.smithy4s     128  thrpt    5     79215.733 ±   1431.124  ops/s
[info] PrimitivesReading.smithy4s        N/A  thrpt    5   1774113.192 ±  10739.137  ops/s
```
After:
```
[info] Benchmark                      (size)   Mode  Cnt         Score        Error  Units
[info] GoogleMapsAPIReading.smithy4s     N/A  thrpt    5     14744.585 ±    310.746  ops/s
[info] NestedStructsReading.smithy4s     128  thrpt    5     95762.699 ±    625.300  ops/s
[info] PrimitivesReading.smithy4s        N/A  thrpt    5   2207861.571 ±  27628.033  ops/s
```